% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDateTimeColumns.R
\name{addDateTimeColumns}
\alias{addDateTimeColumns}
\title{Add date time columns to an antaresData object}
\usage{
addDateTimeColumns(x)
}
\arguments{
\item{x}{An `antaresData` object.}
}
\value{
The function modifies its input by reference. For convenience it returns the
modified input.
}
\description{
This function adds date time columns to an `antaresData` object. It adds a 
time column whose content depends on the time step of the data. It also adds 
other columns depending on the time step.
}
\examples{
opts <- list(
  start = as.POSIXlt("2017-01-01", tz = "UTC"),
  firstWeekday = "Monday"
)

sourcedata <- data.frame(area = "fr", timeId = 1:5, LOAD = rexp(5))
mydata <- as.antaresDataTable(sourcedata, timeStep = "hourly", synthesis = TRUE,
                              type = "areas", opts = opts)

addDateTimeColumns(mydata)
mydata

# Same except we use a daily time
mydata <- as.antaresDataTable(sourcedata, timeStep = "daily", synthesis = TRUE,
                              type = "areas", opts = opts)
addDateTimeColumns(mydata)
mydata
}
