% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateResult.R
\name{aggregateResult_old}
\alias{aggregateResult_old}
\alias{parAggregateMCall_old}
\title{Creation of Mc_all (only antares > V6)}
\usage{
aggregateResult_old(
  opts,
  verbose = 1,
  filtering = FALSE,
  selected = NULL,
  timestep = c("annual", "daily", "hourly", "monthly", "weekly"),
  writeOutput = FALSE,
  mcWeights = NULL,
  mcYears = NULL
)

parAggregateMCall_old(
  opts,
  nbcl = 8,
  verbose = 1,
  timestep = c("annual", "daily", "hourly", "monthly", "weekly"),
  writeOutput = TRUE,
  mcWeights = NULL,
  mcYears = NULL
)
}
\arguments{
\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}}

\item{verbose}{\code{numeric} see logs (1) or not (0)}

\item{filtering}{\code{boolean} filtering control}

\item{selected}{\code{list} named list (pass to antaresRead) : list(areas = 'a', links = 'a - e')}

\item{timestep}{\code{character} antares timestep}

\item{writeOutput}{\code{boolean} write result or not.}

\item{mcWeights}{\code{numeric} vector of weigth for mcYears.}

\item{mcYears}{\code{numeric} mcYears to load.}

\item{nbcl}{\code{numeric} Number of parralel process}
}
\description{
Creation of Mc_all (only antares > V6)
}
\examples{
\dontrun{
   aggregateResult(opts)

}

}
