% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateResult.R
\name{aggregateResult}
\alias{aggregateResult}
\title{Creation of Mc_all}
\usage{
aggregateResult(
  opts,
  verbose = 1,
  filtering = FALSE,
  selected = NULL,
  timestep = c("annual", "daily", "hourly", "monthly", "weekly"),
  writeOutput = TRUE,
  mcWeights = NULL,
  mcYears = NULL
)
}
\arguments{
\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}}

\item{verbose}{\code{numeric} show log in console. Defaut to 1
\itemize{
 \item{0}{ : No log}
 \item{1}{ : Short log}
 \item{2}{ : Long log}
}}

\item{filtering}{\code{boolean} filtering control}

\item{selected}{\code{list} named list (pass to antaresRead) : list(areas = 'a', links = 'a - e')}

\item{timestep}{\code{character} antares timestep}

\item{writeOutput}{\code{boolean} write result or not.}

\item{mcWeights}{\code{numeric} vector of weigth for mcYears.}

\item{mcYears}{\code{numeric} mcYears to load.}
}
\description{
Creation of Mc_all
}
