% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antaresRead-package.R
\docType{package}
\name{antaresRead-package}
\alias{antaresRead}
\alias{antaresRead-package}
\title{antaresRead: Import, Manipulate and Explore the Results of an 'Antares' Simulation}
\description{
Import, manipulate and explore results generated by 'Antares', a powerful open source software developed by RTE (Réseau de Transport d’Électricité) to simulate and study electric power systems (more information about 'Antares' here : \url{https://antares-simulator.org/}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rte-antares-rpackage/antaresRead}
  \item \url{https://rte-antares-rpackage.github.io/antaresRead/}
  \item Report bugs at \url{https://github.com/rte-antares-rpackage/antaresRead/issues}
}

}
\author{
\strong{Maintainer}: Tatiana Vargas \email{tatiana.vargas@rte-france.com}

Authors:
\itemize{
  \item Tatiana Vargas \email{tatiana.vargas@rte-france.com}
  \item Jalal-Edine ZAWAM
  \item Francois Guillem
  \item Benoit Thieurmel
  \item Titouan Robert
}

Other contributors:
\itemize{
  \item Frederic Breant [contributor]
  \item Victor Perrier [contributor]
  \item Etienne Sanchez [contributor]
  \item Assil Mansouri [contributor]
  \item Clement Berthet [contributor]
  \item Kamel Kemiha [contributor]
  \item Abdallah Mahoudi [contributor]
  \item Nicolas Boitard [contributor]
  \item RTE [copyright holder]
}

}
\keyword{internal}
