% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asAntaresDataTable.R
\name{as.antaresDataTable}
\alias{as.antaresDataTable}
\alias{as.antaresDataTable.data.frame}
\title{Convert objects to antaresDataTable}
\usage{
as.antaresDataTable(x, ...)

\method{as.antaresDataTable}{data.frame}(x, synthesis, timeStep, type, opts = simOptions(), ...)
}
\arguments{
\item{x}{object to convert to a an \code{antaresDataList}.}

\item{...}{Arguments to be passed to methods.}

\item{synthesis}{Does the table contain synthetic results ?}

\item{timeStep}{Time step of the data. One of "hourly", "daily", "weekly", "monthly" or "annual".}

\item{type}{type of data: for instance "areas", "links", "clusters", etc.}

\item{opts}{Simulation options.}
}
\value{
\code{antaresDataTable} object.
}
\description{
This function converts a \code{data.frame} or a \code{data.table} into an
\code{antaresDataTable} object.

An \code{antaresDataTable} is simply a \code{data.table} with additional
attributes recording the time step, the type of data and the simulation
options.
}
