% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5_setSimulationPathH5.R
\name{.getOptionsH5}
\alias{.getOptionsH5}
\title{Get H5 options}
\usage{
.getOptionsH5(path)
}
\arguments{
\item{path}{\code{character} path of h5 file}
}
\description{
Get H5 options
}
\keyword{internal}
