% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSimulationPath.R
\name{.transform_antares_version}
\alias{.transform_antares_version}
\title{Convert the Antares number version}
\usage{
.transform_antares_version(antares_version)
}
\arguments{
\item{antares_version}{``numeric` Antares number version.}
}
\value{
\code{numeric} value according to study version
}
\description{
From V9.0, system version is 9.0 (2 digit for minor) instead of 900
}
\keyword{internal}
