% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5_writeH5ByYear.R
\name{.writeAntaresH5Fun}
\alias{.writeAntaresH5Fun}
\title{Convert antares output to h5 file}
\usage{
.writeAntaresH5Fun(
  path,
  timeSteps,
  opts,
  writeMcAll,
  compress,
  misc,
  thermalAvailabilities,
  hydroStorage,
  hydroStorageMaxPower,
  reserve,
  linkCapacity,
  mustRun,
  thermalModulation,
  removeVirtualAreas,
  storageFlexibility,
  production,
  reassignCosts,
  newCols,
  supressMessages
)
}
\description{
Convert antares output to h5 file
}
\keyword{internal}
