% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeographicTrimming.R
\name{getGeographicTrimming}
\alias{getGeographicTrimming}
\title{Read geographic trimming (filtering) options}
\usage{
getGeographicTrimming(areas = NULL, links = TRUE, opts = simOptions())
}
\arguments{
\item{areas}{Character. vector of areas}

\item{links}{Logical. if TRUE, return filtering options for all links starting from selected areas}

\item{opts}{List. simulation options}
}
\value{
list of filtering options for areas and links
}
\description{
Read geographic trimming (filtering) options
}
