% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIds.R
\name{getIdCols}
\alias{getIdCols}
\title{get Id columns}
\usage{
getIdCols(x = NULL)
}
\arguments{
\item{x}{an AntaresDataTable.}
}
\value{
A character vector containing the name of the id columns of an antaresDataTable
}
\description{
\code{getIdCols} return the id columns of an AntaresDataTable
}
