% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIds.R
\name{getIds}
\alias{getIds}
\title{Get Ids of an AntaresDataTable}
\usage{
getIds(x = NULL)
}
\arguments{
\item{x}{an AntaresDataTable.}
}
\value{
A character vector containg ids of an antaresDataTable
}
\description{
\code{getIds} return ids of an AntaresDataTable
}

