% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLinks.R
\name{getLinks}
\alias{getLinks}
\title{Retrieve links connected to a set of nodes}
\usage{
getLinks(nodes, internalOnly = FALSE)
}
\arguments{
\item{nodes}{vector containing node names.}

\item{internalOnly}{if TRUE, only links that connect two nodes from the list are returned. 
If not, the function may return links that connect a node from the list with 
a node outside the list.}
}
\value{
character vector containing link names.
}
\description{
This function find the name of the links connected to a set of node.
}

