% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5utils.R
\name{isH5Opts}
\alias{isH5Opts}
\title{Test if opts is h5}
\usage{
isH5Opts(opts)
}
\arguments{
\item{opts}{, opts}
}
\description{
Test if the value returned by setSimulationPath() is referring to an h5 file
}
