% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAntares.R
\name{readAntaresAPI}
\alias{readAntaresAPI}
\title{Retrieve aggregated areas or links raw data from study output}
\usage{
readAntaresAPI(
  areas = NULL,
  links = NULL,
  select = NULL,
  mcYears = NULL,
  query_file,
  timeStep,
  synthesis = TRUE,
  opts = simOptions()
)
}
\arguments{
\item{areas}{Vector containing the names of the areas to import. If
\code{NULL} all areas are imported.}

\item{links}{Vector containing the name of links to import. If \code{NULL} all
links are imported.}

\item{select}{Names of the columns to be selected.}

\item{mcYears}{Index of the Monte-Carlo years to import. If \code{NULL} all
Monte-Carlo years are imported.}

\item{query_file}{Type of the data to import between
values, details, details-res, details-STstorage}

\item{timeStep}{Resolution of the data to import between hourly, daily,
weekly, monthly or annual.}

\item{synthesis}{Shoudl the query use mc-all or mc-ind.}

\item{opts}{List of simulation parameters returned by the function
\code{\link[=setSimulationPath]{setSimulationPath()}}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
Retrieve aggregated areas or links raw data from study output
}
