% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBindingConstraints.R
\name{readBindingContraints}
\alias{readBindingContraints}
\title{Read binding constraints}
\usage{
readBindingContraints(opts = getOption("antares"))
}
\arguments{
\item{opts}{list of simulation parameters returned by the function 
\code{\link{setSimulationPath}}}
}
\value{
A list with an element per binding constraint.
}
\description{
This function reads the binding constraints of an Antares project.
}

