% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{readDigestFile}
\alias{readDigestFile}
\title{Read digest file}
\usage{
readDigestFile(opts, endpoint = "mc-all/grid/digest.txt")
}
\arguments{
\item{opts}{simulation options}

\item{endpoint}{Suffix of path for digest file
Default is : "mc-all/grid/digest.txt" added to opts$simDataPath}
}
\value{
list of 5 tables (begin, areas, middle, links lin., links quad.)
}
\description{
Read digest file
}
