% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInput.R
\name{readInputTS}
\alias{readInputTS}
\title{Read Input time series}
\usage{
readInputTS(load = NULL, thermalAvailabilities = NULL, ror = NULL,
  hydroStorage = NULL, hydroStorageMaxPower = NULL, wind = NULL,
  solar = NULL, misc = NULL, reserve = NULL, linkCapacity = NULL,
  opts = simOptions(), timeStep = c("hourly", "daily", "weekly", "monthly",
  "annual"), simplify = TRUE, parallel = FALSE, showProgress = TRUE)
}
\description{
Read Input time series
}

