% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLayout.R
\name{readLayout}
\alias{readLayout}
\title{Read nodes layout from the HMI}
\usage{
readLayout(opts = getOption("antares"))
}
\arguments{
\item{opts}{list of simulation parameters returned by the function 
\code{\link{setSimulationPath}}}
}
\value{
A list with two elements
\itemize{
   \item{nodes: }{A data.frame containing the name, the color and the coordinate
                of the nodes}
   \item{links: }{A data.frame containing the name, the coordinates of the origin
                and the destination of the link}
}
}
\description{
This function imports the current node layout from the human machine interface (HMI).
It is useful for plotting the network.
}
\examples{
readLayout()

}

