% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_optim_criteria.R
\name{readOptimCriteria}
\alias{readOptimCriteria}
\title{Read Optimization Criteria}
\usage{
readOptimCriteria(opts = simOptions)
}
\arguments{
\item{opts}{list of simulation parameters returned by the function
\code{\link{setSimulationPath}}}
}
\description{
Read Optimization Criteria
}
