% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setHvdcAreas.R
\name{setHvdcAreas}
\alias{setHvdcAreas}
\title{Set hvdc areas}
\usage{
setHvdcAreas(data, areas)
}
\arguments{
\item{data}{\code{antaresData} or \code{antaresDatalist} data.}

\item{areas}{\code{character} hvdc areas list.}
}
\value{
Object of class "antaresDataList" is returned.
It is a list of data.tables, each element representing one type of element (areas, links, clusters)
}
\description{
This function add hvdc attribute
}
\examples{
\dontrun{

library(antaresRead)
opts <- setSimulationPath('mypath', 1)
myAreaOutput <- readAntares(areas = "all", links = "all")
myAreaOutput <- setHvdcAreas(myAreaOutput, "y_dsr")


}

}
