% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giveSize.R
\name{setRam}
\alias{setRam}
\title{Specify RAM limit}
\usage{
setRam(x)
}
\arguments{
\item{x}{\code{numeric} RAM limit in Go}
}
\description{
This function specify RAM limit (in Go)
}
\examples{
\dontrun{
#Set maximum ram to used to 50 Go
setRam(50)
}

}
