% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_api.R
\name{setSimulationPathAPI}
\alias{setSimulationPathAPI}
\title{Set simualtion path for API}
\usage{
setSimulationPathAPI(path, token, simulation = NULL)
}
\arguments{
\item{path}{\code{character} see setSimulationPath help}

\item{simulation}{\code{character, numeric} see setSimulationPath help}
}
\description{
\code{setSimulationPathAPI} see setSimulationPath function
}
