% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBindingConstraints.R
\name{summary.bindingConstraints}
\alias{summary.bindingConstraints}
\title{Display equation of binding constraint}
\usage{
\method{summary}{bindingConstraints}(object, ...)
}
\arguments{
\item{object}{Object returned by readBindingConstraints}

\item{...}{Unused}
}
\value{
A data.frame with one line per constraint.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function cannot be used for a study \verb{>= 8.7.0}
}
