% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antaresDataAttributes.R
\name{timeStep}
\alias{timeStep}
\alias{setTimeStep}
\alias{synthesis}
\alias{synthesis<-}
\alias{type}
\alias{type<-}
\alias{simOptions}
\alias{simOptions<-}
\title{Get and set attributes of antaresData objects}
\usage{
timeStep(x)

setTimeStep(x, value = c("hourly", "daily", "weekly", "monthly", "annual"))

synthesis(x)

synthesis(x) <- value

type(x)

type(x) <- value

simOptions(x = NULL)

simOptions(x) <- value
}
\arguments{
\item{x}{Object of class `antaresData`. For `type`, it can only  be an
`antaresDataTable` object.}

\item{value}{New value of the attribute.}
}
\value{
Getter functions return the value of the corresponding attribute.
Setter functions are used for their side effect.
}
\description{
These functions permit to get and set attributes of an `antaresData` object.
Four attribtues can be accessed and modified: time step, synthesis, type of
data (areas, links, etc.) and simulation options.
}
\examples{
require(data.table)

mydata <- data.table(area = "fr", timeId = 1:5, LOAD = 1:5)
mydata <- as.antaresDataTable(mydata, synthesis = TRUE, timeStep = "hourly",
                              type = "areas", opts = list())

timeStep(mydata)

# Modify timeStep
timeStep(mydata) <- "daily"
timeStep(mydata)

# It also works on antaresDataList objects
mydata <- as.antaresDataList(mydata)

timeStep(mydata) <- "weekly"
timeStep(mydata)
timeStep(mydata$areas)

}
