% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSimulationPath.R
\name{transform_antares_version}
\alias{transform_antares_version}
\title{Convert the Antares number version}
\usage{
transform_antares_version(antares_version)
}
\arguments{
\item{antares_version}{Antares number version.}
}
\description{
Convert the Antares number version
}
