% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewAntares.R
\name{viewAntares}
\alias{viewAntares}
\title{View the content of an antares output}
\usage{
viewAntares(x)
}
\arguments{
\item{x}{An object of class \code{antaresOutput}, generated by the function
\code{\link{importOutput}}.}
}
\value{
Invisible NULL.
}
\description{
This function displays each element of an \code{antaresOutput} object in a
spreadsheet-like viewer
}

