% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{writeDigest}
\alias{writeDigest}
\title{Write digest file}
\usage{
writeDigest(digest, opts = simOptions())
}
\arguments{
\item{digest}{list of 5 elements similar to what is returned by \code{\link{readDigestFile}}}

\item{opts}{simulation options}
}
\value{
updated digest
list of 5 tables (begin, areas, middle, links lin., links quad.)
}
\description{
Write digest file
}
\seealso{
\code{\link{readDigestFile}}
}
