% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API-methods.R
\name{API-methods}
\alias{API-methods}
\alias{api_get}
\alias{api_post}
\alias{api_put}
\alias{api_delete}
\title{API methods}
\usage{
api_get(
  opts,
  endpoint,
  ...,
  default_endpoint = "v1/studies",
  parse_result = NULL,
  encoding = NULL
)

api_post(opts, endpoint, ..., default_endpoint = "v1/studies")

api_put(opts, endpoint, ..., default_endpoint = "v1/studies")

api_delete(opts, endpoint, ..., default_endpoint = "v1/studies")
}
\arguments{
\item{opts}{Antares simulation options or a \code{list} with an \verb{host = } slot.}

\item{endpoint}{API endpoint to interrogate, it will be added after \code{default_endpoint}.
Can be a full URL (by wrapping ìn \code{\link[=I]{I()}}), in that case \code{default_endpoint} is ignored.}

\item{...}{Additional arguments passed to API method.}

\item{default_endpoint}{Default endpoint to use.}

\item{parse_result}{\code{character} options for parameter \code{as} of function \code{\link[httr:content]{httr::content()}}}

\item{encoding}{argument to pass as argument to the function \code{\link[httr:content]{httr::content()}}}
}
\value{
Response from the API.
}
\description{
API methods
}
\examples{
\dontrun{

# List studies with local API
# default result content in R object (auto parsed)
api_get(opts = list(host = "http://0.0.0.0:8080"),
        endpoint = NULL, 
        parse_result = NULL)

# You can force parse options as text and encoding to UTF-8
api_get(opts = list(host = "http://0.0.0.0:8080"),
        endpoint = NULL, 
        parse_result = "text",
        encoding = "UTF-8")
        
# You can change or delete `default_endpoint` 

# no use `default_endpoint`
api_get(opts = list(host = "http://0.0.0.0:8080"),
        endpoint = NULL, 
        default_endpoint = NULL,
        parse_result = "text",
        encoding = "UTF-8")
        
# replace `default_endpoint`
api_get(opts = list(host = "http://0.0.0.0:8080"),
        endpoint = NULL, 
        default_endpoint = "myfolder/myfolder",
        parse_result = "text",
        encoding = "UTF-8")       
}
}
