% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeTimeStep.R
\name{changeTimeStep}
\alias{changeTimeStep}
\title{Change the timestep of an output}
\usage{
changeTimeStep(x, newTimeStep, oldTimeStep, fun = c("sum", "mean"),
  opts = getOption("antares"))
}
\arguments{
\item{x}{data.table with a column "timeId" or an object of class "antaresOutput"}

\item{newTimeStep}{Desired time step.}

\item{oldTimeStep}{Current time step of the data. If the data has not been manipulated, this 
argument is optional since the time step is stored in an attribute of the
data.}

\item{fun}{function to use for aggregation. Either "sum" or "mean"}
}
\description{
This function changes the timestep of a table or an antaresOutput object
and performs the required aggregation
}

