% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importOutput.R
\name{.download_api_aggregate_result}
\alias{.download_api_aggregate_result}
\title{Retrieve information on a file's state of preparation and retrieve download file.}
\usage{
.download_api_aggregate_result(download_id, opts)
}
\arguments{
\item{download_id}{the id of the download.}

\item{opts}{List of simulation parameters returned by the function
\code{\link[=setSimulationPath]{setSimulationPath()}}}
}
\value{
An updated list containing various information about the simulation.

a data.table
}
\description{
Retrieve information on a file's state of preparation and retrieve download file.
}
