% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBindingConstraints.R
\name{.filter_bindingConstraints_by_names}
\alias{.filter_bindingConstraints_by_names}
\title{Filter a list of binding constraints by names by a exact match}
\usage{
.filter_bindingConstraints_by_names(bindingConstraints, constraint_names)
}
\arguments{
\item{bindingConstraints}{\code{list} a list of binding constraints}

\item{constraint_names}{\code{str} constraint names to filter on}
}
\value{
A list of binding constraints containing the constraints which have their name in constraint_names.
}
\description{
Filter a list of binding constraints by names by a exact match
}
