% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLinks.R
\name{getLinks}
\alias{getLinks}
\title{Retrieve links connected to a set of areas}
\usage{
getLinks(..., opts = simOptions(), internalOnly = FALSE, areas = NULL)
}
\arguments{
\item{...}{parameters passed to \code{\link{getAreas}} to construct a set of nodes.
These parameters are used only if argument \code{areas} is \code{NULL}.}

\item{opts}{list of simulation parameters returned by the function 
\code{\link{setSimulationPath}}}

\item{internalOnly}{If TRUE, only links that connect two areas from the list are returned. 
If not, the function may return links that connect a area from the list with 
a area outside the list.}

\item{areas}{Vector containing area names. It represents the set of areas we interested
in. If \code{NULL}, \code{\link{getAreas}} to create the set of areas.}
}
\value{
character vector containing link names.
}
\description{
This function find the name of the links connected to a set of areas.
}

