% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNodes.R
\name{getNodes}
\alias{getNodes}
\title{Select and exclude nodes}
\usage{
getNodes(select = NULL, exclude = NULL, regexpSelect = TRUE,
  regexpExclude = TRUE, opts = getOption("antares"))
}
\arguments{
\item{select}{Character vector. If \code{regexpSelect} is TRUE, this vector is
interpreted as a list of regular expressions. Else it is interpreted as a
list of node names. If \code{NULL}, all nodes are selected}

\item{exclude}{Character vector. If \code{regexpExclude} is TRUE, this vector is
interpreted as a list of regular expressions and each node validating one
of them is excluded. Else it is interpreted as list of node names to 
exclude. If \code{NULL}, not any node is excluded.}

\item{regexpSelect}{Is \code{select} a list of regular expressions ?}

\item{regexpExclude}{Is \code{exclude} a list of regular expressions ?}

\item{opts}{list of simulation parameters returned by the function 
\code{\link{setSimulationPath}}}
}
\value{
A character vector containg the name of the nodes satisfying the rules
defined by the parameters.
}
\description{
\code{getNodes} is a utility function that builds a list of nodes by using 
regular expressions to select and/or exclude nodes
}

