% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvdcModification.R
\name{hvdcModification}
\alias{hvdcModification}
\title{hvdc straitement}
\usage{
hvdcModification(data, removeHvdcAreas = TRUE, reafectLinks = FALSE)
}
\arguments{
\item{data}{\code{antaresDataList} data to apply straitement}

\item{removeHvdcAreas}{\code{boolean} remove HVDC areas.}

\item{reafectLinks}{\code{boolean} .}
}
\value{
Object of class "antaresDataList" is returned.
It is a list of data.tables, each element representing one type of element (areas, links, clusters)
}
\description{
usage for hvdc
}
\examples{
\dontrun{

data <- readAntares(areas = 'all', links = 'all')
data <- setHvdcAreas(data, "psp in")
data <- hvdcModification(data)

}

}
