% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ponderateMcAggregation.R
\name{ponderateMcAggregation}
\alias{ponderateMcAggregation}
\title{Mcyear aggregation weigthed by wd}
\usage{
ponderateMcAggregation(x, fun = weighted.mean, ...)
}
\arguments{
\item{x}{\code{antaresData} data import with antaresRead}

\item{fun}{\code{function} function to use}

\item{...}{\code{args} others args pass to fun}
}
\value{
Object of class "antaresDataTable".
}
\description{
Mcyear aggregation weigthed by wd
}
\examples{
\dontrun{
  data <- readAntares(areas = 'all', mcYears = 'all')
  ponderateMcAggregation(data, fun = weighted.mean, w = c(.1, .9))

  
}

}
