% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAntares.R
\name{readAntaresNodes}
\alias{readAntaresNodes}
\title{Read output for a list of nodes}
\usage{
readAntaresNodes(nodes, links = TRUE, clusters = TRUE, misc = FALSE,
  thermalAvailabilities = FALSE, hydroStorage = FALSE,
  hydroStorageMaxPower = FALSE, reserve = FALSE, linkCapacity = FALSE,
  internalLinksOnly = FALSE, ...)
}
\arguments{
\item{links}{should links connected to the nodes be imported ?}

\item{clusters}{should the clusters of the nodes be imported ?}

\item{misc}{should misc generation of the nodes be imported ?}

\item{thermalAvailabilities}{Should the thermal availabilities of the nodes be imported ?}

\item{hydroStorage}{Should hydro storage of the nodes be imported ?}

\item{hydroStorageMaxPower}{Should the max power of hydro storage of the nodes be imported ?}

\item{reserve}{Should the reserve of the nodes be imported ?}

\item{linkCapacity}{Should the capacity of the links connected to the nodes be imported ?}

\item{internalLinksOnly}{If FALSE, all links connected to one of the nodes is imported, else
only links connecting two nodes of the list are imported}

\item{...}{Other arguments passed to the function \code{\link{readAntares}}}
}
\value{
If \code{simplify = TRUE} and only one type of output is imported
then the result is a data.table.

Else an object of class "antaresOutput" is returned. It is a list of
data.tables, each element representing one type of element (nodes, links,
clusters)
}
\description{
This a function is a wrapper for "antaresOutput" that reads all data for a
list of nodes.
}
\examples{
\dontrun{
mynode <- getOption("antares")$nodeList[1]
data <- readAntaresNodes(mynode)

# Equivalent but more concise than:
data2 <- readAntares(mynode, links = getLinks(mynode), clusters = mynode)

all.equal(data, data2)
} 
 
}

