% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readClusterDesc.R
\name{readClusterDesc}
\alias{readClusterDesc}
\title{Import clusters description}
\usage{
readClusterDesc()
}
\value{
A data.table with one line per cluster. The columns of the data.table may
change between different projects, but there will always be a column
"cluster" containing the name of the cluster and a column "node" containing the
name of the node it belongs to.
}
\description{
This function creates a table containing description of the clusters of an
antares study: type, capacity, marginal cost, etc.
}
\examples{
# readClusterDesc()

}

