% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeVirtualNodes.R
\name{removeVirtualNodes}
\alias{removeVirtualNodes}
\title{Remove virtual nodes}
\usage{
removeVirtualNodes(x, storageFlexibility = NULL, production = NULL,
  opts = getOption("antares"))
}
\arguments{
\item{x}{An object of class \code{antaresOutput} with at least components 
\code{nodes} and \code{links}}

\item{storageFlexibility}{A vector containing the names of the virtual storage/flexibility nodes}

\item{production}{A vector containing the names of the virtual production nodes}

\item{opts}{list of simulation parameters returned by the function 
\code{\link{setSimulationPath}}}
}
\value{
an \code{antaresOutput object} that has been corrected according to the
specifications of spec.
}
\description{
This function removes virtual nodes from an \code{antaresOutput} object and
corrects the data for the real nodes. The \code{antaresOutput} object should
contain node and link data to function correctly.
}
\examples{
\dontrun{
data <- readAntares(nodes="all", links="all")

# Remove pumped storage virtual nodes

correctedData <- removeVirtualNodes(data, storageFlexibility = c("psp in", "psp out"))
}

}

