% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSimulationPath .R
\name{setSimulationPath}
\alias{setSimulationPath}
\title{Set Path to an Antares simulation}
\usage{
setSimulationPath(path, trace = 1)
}
\arguments{
\item{path}{(optional) Path to the simulation. If missing, a window opens and
lets the user choose the directory of the simulation interactively}

\item{trace}{Level of information to print: 0 for no output, 1 for basic output, 2
for list of nodes and links}
}
\value{
A list containing various information about the simulation
}
\description{
This function has to be used before the import functions. It sets the path to
the Antares simulation and displays information about the simulation
}
\examples{
opts <- setSimulationPath()

}

