% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOptions.R
\name{simOptions}
\alias{simOptions}
\title{Extract the simulation options from an object}
\usage{
simOptions(x)
}
\arguments{
\item{x}{object of class \code{antaresTable} or \code{antaresData}}
}
\value{
list of options used to read the data contained in an object.
}
\description{
The function \code{\link{readAntares}} stores in its output the options used
to read some data (path of the study, node list, link list, start date,
etc.).
}
\details{
\code{simOptions} extracts these options from an object of class 
\code{antaresTable} or \code{antaresOutput}. It can be useful when working on
multiple simulations, either to check how some object has been created or to
use it in some functions like \code{\link{getNodes}} or
\code{\link{getLinks}}
}
\examples{
\dontrun{
  setSimulationPath(study1)
  
  data <- readAntares()
  
  # Choose a different study
  setSimulationPath(study2)
  
  getNodes() # returns the nodes of the secund study
  getNodes(opts = simOptions(data)) # returns the nodes of the first study
  
}

}

