% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewAntares.R
\name{viewAntares}
\alias{viewAntares}
\title{View the content of an antares output}
\usage{
viewAntares(x, ...)
}
\arguments{
\item{x}{An object of class \code{antaresData}, generated by the function
\code{\link{readAntares}}.}

\item{...}{Currently unused}
}
\value{
Invisible NULL.
}
\description{
This function displays each element of an \code{antaresData} object in a
spreadsheet-like viewer.
}
\examples{
\dontrun{
setSimulationPath()

areas <-readAntares()
viewAntares(areas)

output <- studyAntares(areas="all", links = "all", clusters = "all")
viewAntares(output) # Opens three data viewers for each element of output
}

}
